Test1 Finished with the unit tests, covered the code with comments, performed some improvements and fixes.

+++ app/models/Merchant.php
@@ -15,31 +15,68 @@ use Library\Model\BaseDbModel;
 
 class Merchant extends BaseDbModel
 {
+	/**
+	 * The table name.
+	 *
+	 * @var string
+	 */
 	protected static $tableName = 'merchants';
 
+	/**
+	 * The model fields list.
+	 *
+	 * @var array
+	 */
 	protected static $tableFields = [
 		'id'   => 'id',
 		'name' => 'name',
 	];
 
+	/**
+	 * The model relations parameters.
+	 *
+	 * @var array
+	 */
 	protected $relationsMap = [
 		'transactions' => [
 			'relationTable' => 'transactions',
-			'relationModel' => '\App\Models\TransactionTable',
+			'relationModel' => 'App\Models\TransactionTable',
 			'relationKey'   => 'id',
 			'foreignKey'    => 'merchant_id',
 		],
 	];
 
+	/**
+	 * The entry ID.
+	 *
+	 * @var integer
+	 */
 	protected $id;
 
+	/**
+	 * Merchant name.
+	 *
+	 * @var string
+	 */
 	protected $name;
 
+	/**
+	 * Returns the ID.
+	 *
+	 * @return integer
+	 */
 	public function getId()
 	{
 		return $this->id;
 	}
 
+	/**
+	 * Sets the ID.
+	 *
+	 * @param integer $id ID.
+	 *
+	 * @return Merchant
+	 */
 	public function setId($id)
 	{
 		$this->id = $id;
@@ -47,11 +84,23 @@ class Merchant extends BaseDbModel
 		return $this;
 	}
 
+	/**
+	 * Returns the name field value.
+	 *
+	 * @return string
+	 */
 	public function getName()
 	{
 		return $this->name;
 	}
 
+	/**
+	 * Sets the name field value.
+	 *
+	 * @param string $name Name.
+	 *
+	 * @return Merchant
+	 */
 	public function setName($name)
 	{
 		$this->name = $name;
@@ -59,22 +108,41 @@ class Merchant extends BaseDbModel
 		return $this;
 	}
 
+	/**
+	 * Adds a related transaction.
+	 *
+	 * @param TransactionTable $transaction Transaction.
+	 *
+	 * @return void
+	 */
 	public function addTransaction(TransactionTable $transaction)
 	{
 		$transaction->setMerchant($this);
-		$this->relations['transactions'][] = $transaction;
 	}
 
+	/**
+	 * Sets the transactions list and removes previously related.
+	 *
+	 * @param array $transactions Transaction list.
+	 *
+	 * @return void
+	 */
 	public function setTransactions(array $transactions = [])
 	{
-		foreach($this->transactions as $transactions) {
-			$transactions->delete();
+		if (!empty($this->relations['transactions'])) {
+			foreach ($this->relations['transactions'] as $transactions) {
+				$transactions->delete();
+			}
+
+			$this->relations['transactions'] = [];
 		}
-		$this->relations['transactions'] = [];
 
 		foreach ($transactions as $transaction) {
+			if (!$transactions instanceof TransactionTable) {
+				continue;
+			}
+
 			$transaction->setMerchant($this);
-			$this->relations['transactions'][] = $transaction;
         }
 	}
